refactor: Move APIClient to pyinv.tests
This is so that we can use it across multiple apps sensibly
