+++ pyinv/assets/tests/api/test_manufacturers.py
@@ -4,7 +4,7 @@ import pytest
 from django.contrib.auth.models import User
 
 from assets.models import Manufacturer
-from assets.tests.api.client import Client
+from pyinv.tests.client import Client
 
 from .base import APITestCase
