+++ pyinv/assets/tests/api/test_nodes.py
@@ -6,7 +6,7 @@ import pytest
 from assets.models import Asset, Node
 
 from .base import APITestCase
-from .client import Client
+from pyinv.tests.client import Client
 
 
 @pytest.mark.django_db
